backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.132.0",
    "osrswiki-images>=0.1.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.uv.sources]
osrswiki-images = { path = "../osrswiki_images", editable = true }

[tool.pytest.ini_options]
markers = [
    "db: requires a live DATABASE_URL and performs database connectivity checks",
    "live: requires a running backend server and performs live HTTP checks",
]
